+++ changelog.md
@@ -1,7 +1,13 @@
 create the excel export template
 list of transaccion per person
 Import file by company
-Refactor transaction database, amount removed, credit and debit added
+
+
+# v0.0.2: 15.04.2021
+* Response menu fixed
+* Welcome page fixed
+* Remove get_file_content from controller moved to livewire controller
+* Refactor transaction database, amount removed, credit and debit added
 
 # v0.0.1: 12.04.2021
 * get json with php not http:client